src/app/jhp/angular-oidc/services/auth-guard.service.ts
import { Injectable } from '@angular/core';
import { CanActivate, RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
import { AuthService } from './auth.service';
import { Observable } from 'rxjs/Observable';
import 'rxjs/add/operator/map';

@Injectable()
export class AuthGuardService implements CanActivate {
    constructor(private authService: AuthService) {}

    canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
        return this.authService.isLoggedIn().map(isLoggedIn => {
            if (isLoggedIn) {
                return true;
            }

            this.authService.startAuthentication({ state: { target: state.url }});
            return false;
        });
    }
}
